requirements.txt
fastapi
aiohttp
uvicorn[standard]
python-dotenv
aiogram==3.3.0
openai
aiocryptopay
